+++ .snyk
@@ -1,9 +1,9 @@
 # Snyk (https://snyk.io) policy file, patches or ignores known vulnerabilities.
-version: v1.12.0
+version: v1.13.1
 # ignores vulnerabilities until expiry date; change duration by modifying expiry date
 ignore:
   'npm:qs:20170213':
     - http-server > union > qs:
         reason: 'No Patch or Upgrade available '
-        expires: '2018-11-12T10:52:02.751Z'
+        expires: '2018-12-28T00:04:54.899Z'
 patch: {}
